New quoteEscapeChar preference in CsvPreference
This is meant to support reading data where the quote escape character is not
the RFC-compliant second quote character, but instead a separate escape
character, commonly a backslash '\'.

RFC-compliant row:

    one,"two","three with a "" quote character",four

Row this PR is meant to support:

    one,"two","three with a \" quote character",four

We want to be able to parse both to this resulting list:

    [one, two, three with a " quote character, four]

General definition for the quoteEscapeChar is:
- By default the quoteEscapeChar is set to the quoteChar
- To write an escape char into actual data, double the escape char
- An escape char before normal data results in both chars being passed through
- Continue to write data in normal RFC-compliant

Please see additional unit tests written to support this feature and note that
no existing unit tests needed to be modified to support this new option,
indicating the existing behavior is unchanged.